LBAC-1 Changed the module structure based on the classes
Minor changes
