+++ api/src/main/java/org/openmrs/module/locationbasedaccess/aop/EncounterSearchAdviser.java
@@ -8,7 +8,7 @@
  * graphic logo is a trademark of OpenMRS Inc.
  */
 
-package org.openmrs.module.locationbasedaccess;
+package org.openmrs.module.locationbasedaccess.aop;
 
 import org.aopalliance.aop.Advice;
 import org.aopalliance.intercept.MethodInterceptor;
@@ -19,6 +19,7 @@ import org.openmrs.Encounter;
 import org.openmrs.Location;
 import org.openmrs.api.context.Context;
 import org.openmrs.api.context.Daemon;
+import org.openmrs.module.locationbasedaccess.utils.LocationUtils;
 import org.springframework.aop.Advisor;
 import org.springframework.aop.support.StaticMethodMatcherPointcutAdvisor;
